refactor: reorganize project structure and improve documentation
- Restructure examples to be consolidated in root directory instead of subdirectories
- Move documentation generation from main/ to docs/ directory
- Update Makefile to reflect new documentation structure
- Add comprehensive template for documentation site with enhanced styling
- Update .gitignore to exclude new build artifacts and cache directories
- Consolidate example files for better maintainability and discoverability
