+++ .gitignore
@@ -1,3 +1,9 @@
 .idea
-main/static
+docs/static
 vendor
+_site
+docs/_site
+build
+docs/.jekyll-cache
+.bundles
+.claude
