--- main/template/index.html
@@ -1,126 +1,0 @@
-<html lang="en">
-<head>
-    <title>Scalar-go examples</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
-          integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/styles/default.min.css">
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/highlight.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/languages/go.min.js"></script>
-    <style>
-        body {
-            min-height: 100vh;
-            min-height: -webkit-fill-available;
-        }
-
-        main {
-            height: 100vh;
-            height: -webkit-fill-available;
-            max-height: 100vh;
-            overflow-x: auto;
-            overflow-y: hidden;
-        }
-
-        .hide {
-            display: none;
-        }
-
-        .active {
-            display: block;
-        }
-
-        .nav-link {
-            cursor: pointer;
-        }
-    </style>
-</head>
-<body>
-<main class="d-flex flex-nowrap">
-    <div class="d-flex flex-column flex-shrink-0 p-3 text-bg-dark" style="width: 280px;">
-        <a href="/" class="d-flex align-items-center mb-3 mb-md-0 me-md-auto text-white text-decoration-none">
-            <svg class="bi pe-none me-2" width="40" height="32">
-                <use xlink:href="#bootstrap"></use>
-            </svg>
-            <span class="fs-4">Scalar Go</span>
-        </a>
-        <hr>
-        <ul class="sidebar nav nav-pills flex-column mb-auto">
-            <li class="nav-item">
-                <a href="#" class="nav-link active" aria-current="page">
-                    Home
-                </a>
-            </li>
-        </ul>
-    </div>
-    <div class="workspace d-flex flex-column flex-shrink-0 p-3 flex-fill">
-        <ul class="nav nav-pills">
-            <li class="nav-item">
-                <a class="nav-link active" aria-current="page" href="#" for="code">Code</a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link" href="#" for="preview">Preview</a>
-            </li>
-        </ul>
-        <hr/>
-        <div class="tab-content code hide active">
-            <div class="card">
-                <div class="card-header description p-3"></div>
-                <hr/>
-                <pre class="p-3"><code>This is example</code></pre>
-            </div>
-        </div>
-        <iframe class="tab-content preview flex-fill hide"></iframe>
-    </div>
-</main>
-<script type="application/javascript">
-    const examples = {{ . }};
-
-    function renderExample(example, idx) {
-        const blob = new Blob([example.output], {type: "text/html; charset=utf-8"});
-        document.body.querySelector('iframe').src = URL.createObjectURL(blob);
-        document.querySelector('.workspace .tab-content     code').innerHTML = hljs.highlight(example.code, {language: 'go'}).value;
-        document.querySelector('.workspace .description').innerHTML = example.description;
-        document.body.querySelectorAll('.sidebar .nav-link').forEach((el, i) => {
-            el.classList.remove('active');
-        });
-        document.body.querySelector(`.example_${idx}`).classList.add('active');
-    }
-
-    function addExampleToSidebar(example, idx) {
-        const li = document.createElement('li');
-        li.classList.add('nav-item');
-        const a = document.createElement('a');
-        a.classList.add('nav-link', 'text-white', `example_${idx}`);
-        const span = document.createElement('span');
-        span.innerText = example.name;
-        a.appendChild(span);
-        li.appendChild(a);
-
-        a.onclick = function () {
-            renderExample(example, idx);
-        };
-
-        document.querySelector(".sidebar").appendChild(li);
-    }
-
-    (function () {
-        document.querySelectorAll('.workspace .nav-item .nav-link').forEach((el) => {
-            el.addEventListener('click', () => {
-                document.querySelectorAll('.workspace .nav-item .nav-link').forEach((el) => {
-                    el.classList.remove('active');
-                });
-                el.classList.add('active');
-                document.querySelectorAll('.workspace .tab-content').forEach((el) => {
-                    el.classList.remove('active');
-                });
-                document.querySelector('.workspace .tab-content.' + el.getAttribute('for')).classList.add('active');
-            });
-        });
-
-        for (let i = 0; i < examples.length; i++) {
-            addExampleToSidebar(examples[i], i);
-        }
-        renderExample(examples[0], 0);
-    })();
-</script>
-</body>
-</html>
